requirements.txt
certifi==2017.11.5
requests==2.18.4
todoist-python==8.0.0
pytest